+++ nafas/__main__.py
@@ -2,12 +2,13 @@
 """nafas main."""
 
 import sys
+import webbrowser
 import argparse
 from nafas.functions import description_print, get_input_standard, input_filter
 from nafas.functions import get_program_data, program_description_print, run, clear_screen
-from nafas.functions import load_config
+from nafas.functions import load_config, get_rendered_survey_link
 from nafas.params import NAFAS_VERSION, EXIT_MESSAGE
-from nafas.params import CONFIG_LOAD_ERROR_MESSAGE
+from nafas.params import CONFIG_LOAD_ERROR_MESSAGE, SURVEY_MESSAGE
 from art import tprint
 
 
@@ -52,6 +53,8 @@ def main():
                 clear_screen()
                 program_description_print(program_name, level, program_data)
                 run(program_data, silent=silent_flag)
+                if input(SURVEY_MESSAGE).lower() == "y":
+                    webbrowser.open(get_rendered_survey_link(program_name, level, program_data))
                 INPUTINDEX = str(
                     input("Press [R] to restart or any other key to exit."))
                 if INPUTINDEX.upper() != "R":